refactor(tezos): hoist getTezosEarnFlow helper to live-common
Shared between LWD AccountHeaderManageActions and the Tezos delegation
dashboard. LWM will consume the same helper in a follow-up commit.
